Allow search for address search as well as name.
Change database queries not using primary keys in the where clause to use LIKE instead of =. Add a controllers folder and a utility.php file. Utility.php contains a normalize function to normalize and sanitize user input for the database. This allows queries for, example, Apt. #2 to match database entries of apt 2.