Fix address logic in addClient and date processing in utility.
Reject duplicates in the database and fix URLs for redirects on error. If an address is not given, the clients house ID will be NULL. If an address is partially given, an error is raised. If an address is completely given, it is stored regardless of if the person says theyre homeless.