Change reason explanation to be with client.
Instead of having a unique reason for each explanation, just have the explanation be tied to each person. Add the explanation field to the forms and the controllers, leave it optional. Also ensure that if youre editing a client whose reason is lost job that the unemployment date will show up on page load rather than on blur.