Handle address changing better
If a person registers with the same address as another, the new person will have the same house ID as the old person. When either edits their profile, theyre asked if the old address has people still registered with BCC at it. If so, a new house is created. Otherwise, the existing house is merely updated and both people reflect the address change.